minimized LUT to 1/4 size
Implemented minimization of LUTs with two separate verilog functions. One to update the phase_accumulator input to the LUT (keep it between 0 and 256). The other change the output from the LUT by mirroring when necessary.
I tried testing to confirm my work but kept getting the same output on the Signal Tap Analyzer. I tried debugging by using constant values as inputs to the MUX instead of the SIN and TRI LUT outputs. The error seems to be in switching between outputs with the MUX. Maybe something is wrong with my GPIO connectors?