fixed error that arose with clear not working on 7-seg displays
after previous fix clr did not consistently work. Corrected this by making the always statement for phase_accum.v and decOutput.v reliant on posedge clr as well as posedge clk. Becuase the pll is fed the same clr, when clr is pressed, it stops the clock to the clr has a minute chance of catching the code at the correct moment. I had to control for any errors that may arise with the phase accumulator counter by encasing the code in an if statement.